Making types visible again to TypeScript compiler.
I basically reverted a previous change in which a constant of the same shape as the module was exported instead of the module itself, which effectively made the type declarations private. I am not sure that change is needed anymore after reverting the exporting of binaryen as a Promise.

See issue #59 for the effect of the previous change.